+++ lib/faction_viewer/faction_viewer_controller.mapper.dart
@@ -188,6 +188,10 @@ class FactionViewerStatePersistedMapper
     opt: true,
     def: 'combatMedium',
   );
+  static bool _$onlyEnabledMods(FactionViewerStatePersisted v) =>
+      v.onlyEnabledMods;
+  static const Field<FactionViewerStatePersisted, bool> _f$onlyEnabledMods =
+      Field('onlyEnabledMods', _$onlyEnabledMods, opt: true, def: false);
 
   @override
   final MappableFields<FactionViewerStatePersisted> fields = const {
@@ -197,6 +201,7 @@ class FactionViewerStatePersistedMapper
     #gallerySortAscending: _f$gallerySortAscending,
     #spawnFactionKey: _f$spawnFactionKey,
     #spawnRole: _f$spawnRole,
+    #onlyEnabledMods: _f$onlyEnabledMods,
   };
 
   static FactionViewerStatePersisted _instantiate(DecodingData data) {
@@ -207,6 +212,7 @@ class FactionViewerStatePersistedMapper
       gallerySortAscending: data.dec(_f$gallerySortAscending),
       spawnFactionKey: data.dec(_f$spawnFactionKey),
       spawnRole: data.dec(_f$spawnRole),
+      onlyEnabledMods: data.dec(_f$onlyEnabledMods),
     );
   }
 
@@ -291,6 +297,7 @@ abstract class FactionViewerStatePersistedCopyWith<
     bool? gallerySortAscending,
     String? spawnFactionKey,
     String? spawnRole,
+    bool? onlyEnabledMods,
   });
   FactionViewerStatePersistedCopyWith<$R2, $In, $Out2> $chain<$R2, $Out2>(
     Then<$Out2, $R2> t,
@@ -322,6 +329,7 @@ class _FactionViewerStatePersistedCopyWithImpl<$R, $Out>
     bool? gallerySortAscending,
     Object? spawnFactionKey = $none,
     String? spawnRole,
+    bool? onlyEnabledMods,
   }) => $apply(
     FieldCopyWithData({
       if (viewMode != null) #viewMode: viewMode,
@@ -331,6 +339,7 @@ class _FactionViewerStatePersistedCopyWithImpl<$R, $Out>
         #gallerySortAscending: gallerySortAscending,
       if (spawnFactionKey != $none) #spawnFactionKey: spawnFactionKey,
       if (spawnRole != null) #spawnRole: spawnRole,
+      if (onlyEnabledMods != null) #onlyEnabledMods: onlyEnabledMods,
     }),
   );
   @override
@@ -348,6 +357,7 @@ class _FactionViewerStatePersistedCopyWithImpl<$R, $Out>
         ),
         spawnFactionKey: data.get(#spawnFactionKey, or: $value.spawnFactionKey),
         spawnRole: data.get(#spawnRole, or: $value.spawnRole),
+        onlyEnabledMods: data.get(#onlyEnabledMods, or: $value.onlyEnabledMods),
       );
 
   @override